pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project]
name = "nerif"
version = "1.4.0"

description = "LLM powered Python"
readme = "README.md"
authors = [
    {name = "Chivier Humber", email = "chivier.humber@outlook.com"}
]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["llm", "ai", "natural-language", "python", "nerif"]
dependencies = [
    "httpx",
    "pydantic>=2.0",
]
requires-python = ">=3.10"

[tool.hatch.build.targets.wheel]
packages = ["src/nerif"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = ["bumpver", "ruff", "pytest", "pytest-asyncio"]
rag = ["numpy"]
image = ["Pillow>=9.0.0"]
native = ["nerif_native"]
viz = ["jinja2>=3.1"]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]

[project.scripts]
nerif = "nerif.cli.main:main"
nerif-compress = "nerif.cli.compress_image:main"

[project.urls]
Homepage = "https://nerif-ai.com"

[tool.bumpver]
current_version = "1.4.0"

version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "release: Bump version {old_version} -> {new_version}"
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']

[tool.ruff]
exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".ipynb_checkpoints",
    ".mypy_cache", ".nox", ".pants.d", ".pyenv", ".pytest_cache", ".pytype",
    ".ruff_cache", ".svn", ".tox", ".venv", ".vscode", "__pypackages__",
    "_build", "buck-out", "build", "dist", "node_modules", "site-packages", "venv",
]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
